scripts/turism_gov_ro/requirements.txt
beautifulsoup4
click
python-magic
datefinder
jinja2
aiohttp[speedups]
pytest